requirements.txt
python-dotenv==1.0.1
langchain==1.2.17
langchain-core==1.3.2
langchain-community==0.4.1
langchain-google-genai==4.2.2
google-generativeai==0.7.2
pydantic==2.13.3
pyyaml==6.0.3
